Update file structure
