+++ Baby Monitor/Source Files/Services/Analytics/New Group/AnalyticsProtocol.swift
@@ -1,0 +1,8 @@
+//
+//  AnalyticsProtocol.swift
+//  Baby Monitor
+
+protocol AnalyticsProtocol {
+    func logScreen(name: String, className: String)
+    func logEvent(_ eventName: String, parameters: [String: Any]?)
+}
